# docs/features/tasks/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Tasks Board API
  version: 0.1.0
  description: |
    A shared kanban board for a small team (3–7 people): three fixed columns
    (To Do → In Progress → Done), cards with a name and an optional free-text
    assignee, drag-and-drop status changes, and a team-managed unpredictable
    public link that lets people outside the team watch the board's live
    state read-only, with no account of any kind. Editing is deliberately
    unauthenticated — see the `security` note below.
  contact:
    name: Task Tracker
    url: https://github.com/task-tracker
...
servers:
  - url: https://api.example.com
    description: Production
  - url: http://localhost:8080
    description: Local

# Deviation from the repo's usual BearerAuth default: this feature has NO
# authentication anywhere, by deliberate spec decision (spec.md §3 Non-goals,
# §6.1) — editing the board requires no login at all, and the public-view
# surface is gated only by possessing an unguessable link token (ADR-0003),
# not by a bearer token. `security: []` at the root applies to every
# operation; the BearerAuth scheme is kept declared below only because the
# repo's platform defines it for other modules, not because this API uses it.
security: []
paths:
...
  /api/v1/public/boards/{token}:
    get:
      summary: Resolve a public link and read the board read-only (US-04, AC-05, AC-06, AC-08)
      description: The token is the sole access control — no session, no cookie, no header. A disabled or unknown token is indistinguishable from any other bad address.
      operationId: getPublicBoard
      tags: [public]
      security: []
      parameters:
        - in: path
          name: token
          required: true
          schema: { type: string }
      responses:
        "200":
          description: >-
            The board's current state, read-only — never a stale snapshot
            (AC-08). The response carries no edit affordance of any kind
            (AC-06); that is a UI rendering property, not a field this
            schema needs to express.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CardPage" }
        "404":
          description: >-
            The link is disabled or was never valid. Deliberately
            byte-for-byte identical to the server's generic not-found for
            any unmatched route (AC-05) — plain text, not this API's usual
            `{code,message}` JSON envelope, since even that structured shape
            would itself confirm a board once existed at this address.
          content:
            text/plain:
              schema: { type: string }
              example: "404 page not found"
  /api/v1/public/boards/{token}/events:
    get:
      summary: Live stream of board changes for a viewer holding an active public link (ADR-0001)
      description: >-
        Server-Sent Events, scoped to one link token. Pushes the same board
        events as the editor stream while the link stays active; pushes a
        `link.disabled` message (or simply closes) the moment the link is
        disabled, which the viewer's page turns into an automatic
        transition to not-found (AC-12) without a manual refresh.
      operationId: subscribePublicBoardEvents
      tags: [public]
      security: []
      parameters:
        - in: path
          name: token
          required: true
          schema: { type: string }
      responses:
        "200":
          description: An open `text/event-stream` connection, scoped to this link.
          content:
            text/event-stream:
              schema: { type: string }
        "404":
          description: The link is disabled or was never valid — same plain-text generic not-found as the read endpoint above.
          content:
            text/plain:
              schema: { type: string }
              example: "404 page not found"
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >-
        Declared for parity with the repo's platform (used by other
        modules, e.g. `user`/`auth`) — no operation in this contract
        references it; every operation is `security: []`.
...
  schemas:
...
    Card:
      type: object
      required: [id, name, column_status, created_at, updated_at]
      additionalProperties: false
      properties:
        id: { type: string, format: uuid }
        name: { type: string, maxLength: 200 }
        assignee: { type: [string, "null"], maxLength: 100 }
        column_status:
          type: string
          enum: [todo, in_progress, done]
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
    CardPage:
      type: object
      required: [items, has_next, has_prev]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/Card" }
        has_next: { type: boolean }
        has_prev: { type: boolean }
        next_cursor: { type: [string, "null"], format: uuid }
...